libct: move criu-related stuff to separate file
No code change, only added periods to some comments to make godot happy.
